Initial commit.
Suddenly this whole thing has come together. Initially I was creating images via RVG, but since that doesn't have any form of text-flowing, it was always going to be a pain. Designing in the browser is very easy, on the otherhand. PhantomJS isn't perfect for rasterising, and I suspect that I could tweak the output image a bit using RMagick, but this demonstrates an end-to-end flow:

1. Submit a page to the server
2. The server creates a job to rasterize the page
3. The rasterisation job creates and image, and then queues a job to turn the image data into printer-friendly byte format
4. The image to bytes jobs processes the image and puts it onto a printer-id-specific queue
5. The printer (which has been polling a URL) picks up the data, and starts printing.

... and it works. Scary.

The Arduino sketch sometimes seizes up (in particular there's a bug if the content length is greater than MAXINT which I haven't tracked down), but to some extend this is the walking skeleton of an open source Wee Printer system.