The Arduino doesn't like talking to shotgun.
For some reason, it would just never connect. Most peculiar! It looks like shotgun listens on the 127.0.0.1 interface only by default, but even using the `-o 0.0.0.0` option (which did allow connections on a 192.168.1.x IP), the Arduino still wouldn't connect.

Ah well. This is simpler for debugging at the moment anyway - change what's going to be printed by editing the view, which is already reloaded.