More robust version of Arduino sketch.
This implements a more comprehensive system for polling, downloading and finally printing an image. Once the polled URL has some content, it is downloaded, and the ready LED is lit. Pressing the button will then trigger printing.

This is as a result of some fairly exhaustive testing on different networks with a stripped down sketch. On my home network, downloads are completely reliable from both local and remote servers.