Don't prevent POST requests for preview or print from other servers.
Web pages should be able to use javascript/ajax to POST their bodies to the backend server for preview. However, Rack::Protection::JsonCsrf (enabled by default in Sinatra) will stop this, because the request Referer header doesn't match the server.

I understand that this opens up a potential CSRF hole, but given that there is no application "session" to be hijacked, I prefer to make the useful functionality possible.