Give better feedback if a preview URL can't be loaded.
I've seen the preview queue get stuck because PhantomJS wouldn't exit if it couldn't resolve the URL for rasterisation. Fixing this was simple (adding an exit to that code path), but it is even better to communicate the problem back to the user.

We could avoid this even earlier by checking the parameters for validity within the controller, but this is a good failsafe.