Improve reliability of first-boot ID generation.
This should fix issue #19.

We've seen a few reports of ID generation failing, resulting in an ID of
the form "i%ff%ff%ff%ff". We inspected a brand new Arduino Ethernet and
found that the first byte of its EEPROM was not 255 as the documentation
suggests, but 105 (ASCII i). This would explain why the ID wasn't
generated.

We could've checked for 105 as the first byte directly, since that's not
a valid first character for an ID (it should be a number), but that felt
more risky. We're also not 100% sure that Arduinos will always have the
first byte set to 105, so this fix should be a bit more permissive of
any other first-byte weirdness that other Arduino models have.