Update the Printer Arduino sketch.
We should always set the mode of the 3 LED pins to OUTPUT, irrespective
of whether we've enabled debugging or not.

I did consider splitting the initDiagnosticLEDs function into two: one
containing the logic for initialising the LED pins and the other to turn
then on for a second. I mentioned this to @lazyatom and he suggested
that it's probably useful to always flash the LEDs at boot time.