Remove the version header as it was causing errors.
When the version header was present, the Arduino would download data, but then
fail because the cache size (the file on the SD) was apparently zero length. I
tried this with other dummy headers, and basically any other header seemed to
screw up the Arduino cache.

Why sending another header would affect the behaviour of the downloading, or
the SD card, I have absolutely no idea :(

I've tried making the requests locally via curl, but the responses are completely identical, even down to a binary diff.