Fail fast if we can't clear the cache.
After initialization, the error LED will flash three times, indicating a
failure to clear the cache.  The sketch will then 'terminate' in as much
as it won't do anything in the main loop.

I've seen the sketch fail to clear the cache on multiple occassions.
I'm not 100% sure how it gets in this state but the only solution
appears to be deleting the cache ('TMP') or formatting the card in
another machine.

There's some duplication between `abortDueToCacheFailure` and
`recoverableError` that I should be able to factor out later.

I've removed the `systemError` function and inlined the switching on of
the error LED.  This is so that I can use a boolean named systemError to
indicate that the sketch should terminate.