Abort if there are problems with the SD Card.
If there are problems opening the SD card for access then there's no
point continuing as we can't cache the data.

Slightly controversially, I've chosen the bump the error codes (number
of LED flashes) of a DHCP failure and cache failure.  I don't think this
is too much of a problem, partcularly because I've bumped the version
number.