Fix problem with crazy large durations..
If the connection to printer.gofreerange.com times out (after about 33
seconds) then we try to report the duration of the request.
Unfortunately, this relies on `start` being assigned a sensible value,
which was only happening after a successful connection.  Prior to this fix,
when the connection timed out, the duration would be reported as a very
large number like 4289767837.

This commit also removes the (in hindsight, obvious) compiler warning,
`printer.cpp:176: warning: 'start' may be used uninitialized in this
function`.