Remove some unnecessary debug statements.
*NOTE* The main motivation is to reduce the size of the data we're
storing in RAM.  The sketch before this commit would fail to print when
debug was enabled.  The image would be downloaded successfully but would
appear not to be cached (cache.size() returned 0).

* Remove 'dhcp fail' debug statement. The flashing error LEDs, along
with the code comment, should provide enough information.

* Remove 'failed to clear cache' debug statement. The flashing error
LEDs, along with the code comment, should provide enough information.

* Remove 'cache cleared' debug statement. The sketch will 'terminate'
if we couldn't clear the cache.  If it doesn't terminate, we can imply
that the cache was cleared.

* Remove 'starting request, cache size is' debug statement.
This statement wasn't adding any useful information.  If our sketch
makes it to the 'Attempting to connect to...' line, then we know that
the cache was 0, and that we're trying to connect.

* Remove 'still connected' debug statement. Both the 'still connected'
and 'waiting for data' debug statements are printed within the loop
that continues all the time the client is connected.  I think that
one of these statements is sufficient.