Store print archives in a less memory-hungry data store.
The Redis instance for storing the printer data was using >800MB of
RAM, because it was storing all prints. This is too much to run on a
reasonable VPS.

I've introduced PostgreSQL as a new data store, and that's where
prints are now persisted. In the future, it's likely that more data
will move there, particularly data that doesn't need to expire.

In order to migrate the data, two scripts are provided:

    bundle exec ruby migrate_prints_to_database.rb

Once this has worked, you can remove the redis archives with:

    bundle exec ruby remove_archive_from_redis.rb